*Kayıt ve silme işlemlerinin kontrolleri yapıldı. *Tasarım değişikliği yapıldı

+++ js/file.js
@@ -1,38 +1,49 @@
 let FILES;
 let FILE_CATALOG;
+let DURUM;
 
 let file = {
     saveData: () => {
+        DURUM = 0;
         minify(); // dosya içeriğinin beautify özelliği geri alınır
-        let filename = $("#filename").val();
+        let filename = (($("#filename").val()).toUpperCase()).trim();
+        for (let i = 0; i <= FILE_CATALOG.length; i++) {
+            if (filename == FILE_CATALOG[i]) {
+                DURUM = 1;
+            }
+        }
         if (!!filename) {
-        	document.getElementById("uyari").innerHTML=" ";
+            //$("#filename").setAttribute("disabled", "");
             let fileData = { filename: filename, file: $("#file").val() };
-            let index = FILE_CATALOG.findIndex(e => e.filename == filename);
+            let index = FILE_CATALOG.findIndex(e => e == filename);
             let isOld = index != -1 ? true : false;
-            isOld ? FILE_CATALOG[FILE_CATALOG.findIndex(e => e.filename == filename)] = fileData : FILE_CATALOG.push(fileData);
-            $.ajax({
-                    url: 'file-api.php',
-                    type: 'POST',
-                    contentType: "application/json; charset=utf-8",
-                    data: JSON.stringify({
-                        fileData: fileData,
-                        catalogData: FILE_CATALOG,
+            isOld ? FILE_CATALOG[index] = fileData : FILE_CATALOG.push(filename);
+            if (DURUM == 0) {
+                $.ajax({
+                        url: 'file-api.php',
+                        type: 'POST',
+                        contentType: "application/json; charset=utf-8",
+                        data: JSON.stringify({
+                            fileData: fileData,
+                            catalogData: FILE_CATALOG,
+                        })
                     })
-                })
-                .done(function() {
-                    updateSlides(fileData, index);
-                    console.log("success");
-                })
-                .fail(function() {
-                    alert("HATA OLUŞTU!")
-                    console.log("error");
-                })
-                .always(function() {
-                    console.log("complete");
-                });
+                    .done(function() {
+                        updateSlides(filename, isOld);
+                        console.log("success");
+                    })
+                    .fail(function() {
+                        alert("HATA OLUŞTU!")
+                        console.log("error");
+                    })
+                    .always(function() {
+                        console.log("complete");
+                    });
+            } else {
+                alert("Bu isimle başka kayıt var!");
+            }
         } else {
-        	document.getElementById("uyari").innerHTML="***Dosya adını boş bırakmayınız...";
+            uyari();
         }
 
     },
@@ -62,27 +73,33 @@ let file = {
     deleteFile: ($arg) => {
         event.stopPropagation();
         let filename = $arg.siblings("div").text();
-        $.ajax({
-                url: 'file-api.php',
-                type: 'DELETE',
-                data: JSON.stringify({
-                    filename: filename
+        var agree = confirm("Bu içeriği silmek istediğinizden emin misiniz?\nBu işlem geri alınamaz!");
+        if (agree) {
+            $.ajax({
+                    url: 'file-api.php',
+                    type: 'DELETE',
+                    data: JSON.stringify({
+                        filename: filename
+                    })
                 })
-            })
-            .done(function(data) {
-                let index = FILE_CATALOG.findIndex(e => e.filename == filename);
-                FILE_CATALOG.splice(index, 1);
-                $arg.parents(".slide").remove();
+                .done(function(data) {
+                    let index = FILE_CATALOG.findIndex(e => e.filename == filename);
+                    FILE_CATALOG.splice(index, 1);
+                    $arg.parents(".slide").remove();
 
-                file.saveOnlyCatalogData(filename);
-                console.log("success");
-            })
-            .fail(function() {
-                console.log("error");
-            })
-            .always(function() {
-                console.log("complete");
-            });
+                    file.saveOnlyCatalogData(filename);
+                    console.log("success");
+                })
+                .fail(function() {
+                    console.log("error");
+                })
+                .always(function() {
+                    console.log("complete");
+                });
+            document.getElementById("filename").value = "";
+            document.getElementsByClassName("trumbowyg-editor")[0].innerHTML = "";
+        } else { return false; }
+
     },
 
     getCatalogData: (ilkMi = false) => {
@@ -136,16 +153,16 @@ formReset();
 $('.trumbowyg').trumbowyg({
     lang: 'tr',
     /*btns: [
-    	['undo', 'redo'], // Only supported in Blink browsers
-    	['formatting'],
-    	['strong', 'em', 'del'],
-    	['superscript', 'subscript'],
-    	['justifyLeft', 'justifyCenter', 'justifyRight', 'justifyFull'],
-    	['unorderedList', 'orderedList'],
-    	['horizontalRule'],
-    	['removeformat'],
-    	['fullscreen'],
-    	['foreColor', 'backColor']
+        ['undo', 'redo'], // Only supported in Blink browsers
+        ['formatting'],
+        ['strong', 'em', 'del'],
+        ['superscript', 'subscript'],
+        ['justifyLeft', 'justifyCenter', 'justifyRight', 'justifyFull'],
+        ['unorderedList', 'orderedList'],
+        ['horizontalRule'],
+        ['removeformat'],
+        ['fullscreen'],
+        ['foreColor', 'backColor']
     ]*/
     btns: [
         ['viewHTML'],
@@ -159,6 +176,7 @@ $('.trumbowyg').trumbowyg({
         ['unorderedList', 'orderedList'],
         ['horizontalRule'],
         ['removeformat'],
+        ['code'],
         ['fullscreen']
     ]
 });
@@ -167,12 +185,44 @@ FILE_CATALOG = file.getCatalogData();
 
 
 function generateSlide(data) {
-    return '<li class="slide leaderboard__profile" onclick="file.getFileData($(this))"><div>' + data.filename +
+    return '<li class="slide leaderboard__profile" onclick="file.getFileData($(this))"><div class="metin_genislik">' + data +
         '</div><span class="btn-delete" onclick="file.deleteFile($(this))">X</span></li>';
 }
 
-function updateSlides(fileData, index) {
-    let isOld = index != -1 ? true : false;
+function updateSlides(filename, isOld) {
+    //let isOld = index != -1 ? true : false;
     if (!isOld)
-        $('#slides').append(generateSlide(fileData));
+        $('#slides').append(generateSlide(filename));
 }
+
+function closeEditorWarning() {
+    return 'Sayfadan ayrılırsanız bilgileriniz kaydedilmeyecektir'
+}
+
+window.onbeforeunload = closeEditorWarning;
+
+$(".trumbowyg-viewHTML-button").click(function(event) {
+    beautify();
+});
+
+function uyari() {
+    $("#filename").addClass("uyari-border");
+}
+
+function uyariKaldir() {
+    $("#filename").removeClass("uyari-border");
+}
+
+/*function codeBlokOlustur() {
+  "use strict";
+ var bolum = document.getElementsByClassName("trumbowyg-editor");
+ bolum.innerHTML = "Merhaba JavaScript";
+ 
+ document.body.appendChild(bolum);}*/
+
+function codeBlokOlustur() {
+    document.getElementsByClassName("trumbowyg-editor")[0].innerHTML = "deneme";
+}
+
+var btnYaz = document.getElementById("btn");
+btnYaz.onclick = codeBlokOlustur;
